--- Benchmarks/OTelTraceIdRatioBasedSamplerBenchmark/OTelTraceIdRatioBasedSamplerBenchmark.swift
@@ -1,28 +1,0 @@
-// Benchmark boilerplate generated by Benchmark
-
-import Benchmark
-import OTel
-import W3CTraceContext
-
-let benchmarks : @Sendable () -> Benchmark? = {
-    Benchmark("OTelTraceIdRatioBasedSampler") { benchmark in
-
-        let sampler = OTelTraceIdRatioBasedSampler(ratio: 0.5)
-
-        let traceIds = benchmark.scaledIterations.map { _ in TraceID.random() }
-        
-        benchmark.startMeasurement()
-
-        for traceId in traceIds {
-            _ = sampler.samplingResult(
-                operationName: "some-op",
-                kind: .internal,
-                traceID: traceId,
-                attributes: [:],
-                links: [],
-                parentContext: .topLevel)
-        }
-
-        benchmark.stopMeasurement()
-    }
-}
