+++ Tests/OTelTests/Tracing/Sampling/OTelTraceIdRatioBasedSamplerTests.swift
@@ -1,3 +1,16 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift OTel open source project
+//
+// Copyright (c) 2023 Moritz Lang and the Swift OTel project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
 @testable import OTel
 import XCTest
 import Tracing
@@ -32,7 +45,6 @@ final class OTelTraceIdRatioBasedSamplerTests: XCTestCase {
 
         XCTAssertEqual(result, OTelSamplingResult(decision: .recordAndSample, attributes: [:]))
     }
-
 
     func test_different_ratios() {
